sol/solution_test.go
package sol

import (
	"reflect"
	"testing"
)

func BenchmarkTest(b *testing.B) {
	digits := "23"
	for idx := 0; idx < b.N; idx++ {
		letterCombinations(digits)
	}
}
func Test_letterCombinations(t *testing.T) {
	type args struct {
		digits string
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{
			name: "23",
			args: args{digits: "23"},
			want: []string{"ad", "ae", "af", "bd", "be", "bf", "cd", "ce", "cf"},
		},
		{
			name: "",
			args: args{digits: ""},
			want: []string{},
		},
		{
			name: "2",
			args: args{digits: "2"},
			want: []string{"a", "b", "c"},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := letterCombinations(tt.args.digits); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("letterCombinations() = %v, want %v", got, tt.want)
			}
		})
	}
}
